Remove functions to construct an empty message.
A 0MQ message with 0 frames is invalid so don't make it easy to
construct one.  Removed the functions to create an empty Message.  The
Message structure has to be created from some other data so that it
has at least one frame.  This won't keep someone from poping all the
frames and having an empty message structure, but it should at least
make it harder to have an empty message on accident.