Add stock ticker example that uses multiframe PUB/SUB.
Server sends out messages with two frames.  The first frame is the
stock symbol AAA, ABB, or BBB with a random price in the second
frame. The client prints out the stock symbol and price.  The client
can take an optional argument that it can use to create a
subscription.  For example:

stock_client A

would get only messages for stocks AAA and ABB.

stock_client AA would get only AAA and so forth.