Make functions that could result in empty message pub(crate)
Made pop_front and pop_back pub(crate) so that you can't accidently
create an empty message by popping off all the Frames.  Also made
creation of a message from a Vec or VecDeque TryFrom instead of From
to make sure and check that an empty container is not used to create
an empty message.