Make all socket types support multipart messages. (#122)
* Make all socket types support multipart messages.

The specification says all messages can contain multiple frames.  This
commit gets rid of the distinction between multipart and non multipart
messages.  It does this by having all user facing methods take and
return ZmqMessage structures.  I also modified that structure to be a
wrapper around a VecDeque of Bytes.

* Remove functions to construct an empty message.

A 0MQ message with 0 frames is invalid so don't make it easy to
construct one.  Removed the functions to create an empty Message.  The
Message structure has to be created from some other data so that it
has at least one frame.  This won't keep someone from poping all the
frames and having an empty message structure, but it should at least
make it harder to have an empty message on accident.

* Add stock ticker example that uses multiframe PUB/SUB.

Server sends out messages with two frames.  The first frame is the
stock symbol AAA, ABB, or BBB with a random price in the second
frame. The client prints out the stock symbol and price.  The client
can take an optional argument that it can use to create a
subscription.  For example:

stock_client A

would get only messages for stocks AAA and ABB.

stock_client AA would get only AAA and so forth.

* Make functions that could result in empty message pub(crate)

Made pop_front and pop_back pub(crate) so that you can't accidently
create an empty message by popping off all the Frames.  Also made
creation of a message from a Vec or VecDeque TryFrom instead of From
to make sure and check that an empty container is not used to create
an empty message.

* Add to_vec and to_vecdeque functions for message

Since we no longer allow popping frames from a ZmqMessage create a way
to copy the frames out of the struct if that kind of manipulation is desired.

* Remove pop_back function

Not used anywhere in the crate yet so remove it until we need it.

* Rust format run

* Have to_vec and to_vecdeque take self as value.

Avoids copying data

* Change functions names to follow Rust conventions.

Make clippy happy.