Added basic SocketOptions struct
This is related to #116.

Currently it's only enables user to set custom peer-identity if needed
and passes options all around. Later we can add other options to this
struct.

Some minor changes to ReadyCommand properties were required cause
ZMQ_IDENTITY could be any byte vec and not guaranteed to be valid String